refactor: close the guard's advisory window and correct the docs
The unreachable-symbol guard was seeded red at 134 symbols and wired as a
non-blocking CI step so the cleanup units could land while it reported. The
sweep brought that to 25; this change resolves the remainder and makes the
guard blocking -- continue-on-error is gone from the CI step and
deadcode-check joins just ci-check.

The 25 were not one thing:

Three .Error() methods on config.FieldValidationError,
config.MultiValidationError and validator.ValidationError satisfy the error
interface and are called through it, exactly like the Stringers already
exempt. They are false positives; deleting them would break error reporting.
Allowlisted as reflection-dispatched.

Nine more are allowlisted as retained public surface. Seven are the
SanitizeStruct reflection chain that GOTCHAS section 14.4 documents as an
opt-in consumer API with a known gap tracked separately -- the CLI uses
SanitizeXML instead. RuleEngine.Redact and ShouldRedactField are cited
separately rather than folded into that group: they are not part of that call
chain, and ShouldRedactField's own doc comment already records that it has no
caller outside tests.

Eleven were genuinely dead with a live sibling already serving the real path,
and are deleted or moved. Where a test carried real behavioral coverage the
test was repointed rather than dropped -- formatters.New to NewWithMode,
NewJSONFormatterCompact to a struct literal. Where a symbol was only ever a
test convenience it moved into a _test.go file, following the rule in
docs/development/standards.md and the precedent this branch set with
test_helpers.go: RiskSummary.HasRisks, sanitizer.ValidModes, and
display.Theme.GetColor, the last newly dead because removing the progress
path took its only caller with it.

The docs sweep found more than the deleted symbols. cli-output-policy.md
prescribed charmbracelet/bubbles/progress as the canonical progress channel
for a feature that never had a production caller; that channel is restated
honestly rather than just losing the dependency name. api.md documented the
deleted Converter interface, pipelines.md a deleted registry method,
xml-structure-research.md two FormatBoolean claims that were never true, and
README.md and public-api.md still credited bubbles after it left go.mod. A
.golangci.yml goconst exclusion still named a deleted file.
