+++ internal/diff/analyzers/normalizer_test.go
@@ -85,47 +85,3 @@ func TestNormalizer_NormalizePort(t *testing.T) {
 		})
 	}
 }
-
-func TestNormalizer_NormalizeProtocol(t *testing.T) {
-	n := NewNormalizer()
-
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		{"lowercase", "tcp", "tcp"},
-		{"uppercase", "TCP", "tcp"},
-		{"mixed case", "Tcp", "tcp"},
-		{"with spaces", " tcp ", "tcp"},
-		{"empty", "", ""},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.expected, n.NormalizeProtocol(tt.input))
-		})
-	}
-}
-
-func TestNormalizer_NormalizePath(t *testing.T) {
-	n := NewNormalizer()
-
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		{"no change", "/etc/config", "/etc/config"},
-		{"trailing slash", "/etc/config/", "/etc/config"},
-		{"double slashes", "/etc//config", "/etc/config"},
-		{"multiple trailing", "/etc/config///", "/etc/config"},
-		{"empty", "", ""},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.expected, n.NormalizePath(tt.input))
-		})
-	}
-}
