+++ internal/diff/security/risk.go
@@ -20,14 +20,6 @@ type RiskItem struct {
 	Impact      string `json:"impact"`
 }
 
-// HasRisks returns true if any security impacts were detected.
-func (r *RiskSummary) HasRisks() bool {
-	if r == nil {
-		return false
-	}
-	return r.High > 0 || r.Medium > 0 || r.Low > 0
-}
-
 // Severity weights for scoring.
 const (
 	weightHigh   = 10
